refactor: MovieInformationDTO 파일 분리
