+++ BoxOffice/DTO/MovieInfoResultDTO.swift
@@ -1,0 +1,18 @@
+//
+//  MovieInfoResultDTO.swift
+//  BoxOffice
+//
+//  Created by 박재우 on 2023/05/01.
+//
+
+import UIKit
+
+struct MovieInfoResultDTO {
+    let movieInformation: MovieInformationDTO
+    let source: String
+
+    enum CodingKeys: String, CodingKey {
+        case movieInformation = "movieInfo"
+        case source
+    }
+}
